Move RowExpressionCanonicalizer out of ExpressionEquivalence
